+++ src/app/blog/page.tsx
@@ -1,3 +1,4 @@
+import { Metadata } from "next";
 import { Suspense } from "react";
 import ClientPage from "./_components/ClientPage";
 
@@ -9,4 +10,27 @@ export default function Page() {
       </Suspense>
     </div>
   );
+}
+
+export async function generateMetadata(): Promise<Metadata> {
+  try {
+    return {
+      title: "블로그 게시글 목록 | JMJ's Devlog",
+      description: "JMJ의 개발 블로그 게시글 목록입니다.",
+      openGraph: {
+        title: "블로그 게시글 목록 | JMJ's Devlog",
+        description: "JMJ의 개발 블로그 게시글 목록입니다.",
+        type: "website",
+      },
+      robots: {
+        index: true,
+        follow: true,
+      },
+    };
+  } catch (error) {
+    return {
+      title: "블로그 포스트",
+      description: "블로그 포스트를 찾을 수 없습니다.",
+    };
+  }
 }
